Removes FrameLayout Fragment in favor of keeping that functionality the only bits of MainActivity2, finally realizes that because I never replaced the Navigation Layout with the Navigation Fragment, it never gained the functionality of that kotlin file, it still seems as if are now using the ViewModel to controll navigation in a way that spans an activity and a fragment, gets App Live again, realized that we lose data peremeance on a rotateScreen
