--- app/src/main/res/layout/fragment_nav.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:Context=".NavFragment">
-
-    <com.google.android.material.bottomappbar.BottomAppBar
-        android:id="@+id/mainAppBar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom"
-        android:background="@color/namespaceBlack"
-        app:fabCradleMargin="16dp"
-        app:fabCradleRoundedCornerRadius="50dp"
-        app:fabCradleVerticalOffset="5dp">
-
-        <com.google.android.material.bottomnavigation.BottomNavigationView
-            android:id="@+id/mainNavigationView"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_marginEnd="16dp"
-            android:background="@android:color/transparent"
-            app:menu="@menu/main_nav_menu">
-
-        </com.google.android.material.bottomnavigation.BottomNavigationView>
-
-    </com.google.android.material.bottomappbar.BottomAppBar>
-
-    <include
-        android:id="@+id/omnifab"
-        layout="@layout/omnifab"
-        android:layout_width="97dp"
-        android:layout_height="wrap_content"
-        app:layout_anchor="@+id/mainNavigationView"
-        app:layout_anchorGravity="center" />
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
